+++ src/Api/TemperatureController.cs
@@ -3,6 +3,7 @@
  */
 
 using Microsoft.AspNetCore.Mvc;
+using UsbtempServer.Thermology;
 
 namespace UsbtempServer.Api;
 
@@ -10,9 +11,9 @@ namespace UsbtempServer.Api;
 [Route("temperature")]
 public class TemperatureController : ControllerBase
 {
-	private readonly Thermometer thermometer;
+	private readonly IThermometer thermometer;
 
-	public TemperatureController(Thermometer thermometer)
+	public TemperatureController(IThermometer thermometer)
 	{
 		this.thermometer = thermometer;
 	}