+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [**Semantic Versioning v2.0.0**](https://semver.org/
 ### Added ###
 
 * Auto-check for available updates on startup ([`8468824`](https://github.com/mfederczuk/usbtemp-server/commit/84688245824f6d22f3fc6cf5f24b9c41bd6dd4a2))
+* (debug builds only) Allow using a virtual mock thermometer instead of a physical one ([`1045013`](https://github.com/mfederczuk/usbtemp-server/commit/1045013124c0f62df9018774136a5d5bb42625f1))
 
 ## [v0.1.0-indev01] - 2023-07-25 ##